+++ src/Ext/WarheadType/Body.h
@@ -278,6 +278,7 @@ public:
 	Valueable<bool> Malicious;
 	Nullable<int> Flash_Duration;
 	Valueable<double> Damage_Deployed { 1.0 };
+	Nullable<bool> PreventScatter;
 
 	double Crit_RandomBuffer;
 	double Crit_CurrentChance;
@@ -513,6 +514,7 @@ public:
 		, EffectsRequireVerses { true }
 		, Malicious { true }
 		, Flash_Duration {}
+		, PreventScatter {}
 
 		, Crit_RandomBuffer { 0.0 }
 		, Crit_CurrentChance { 0.0 }